+++ src/main/java/br/com/alura/orcamentoapi/model/Resumo.java
@@ -3,19 +3,19 @@ package br.com.alura.orcamentoapi.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
-import java.util.Map;
 
 @Getter
 @Setter
 public class Resumo {
 
-    private Float vTotalReceitas;
-    private Float vTotalDespesas;
-    private Float saldoFinal;
+    private BigDecimal vTotalReceitas;
+    private BigDecimal vTotalDespesas;
+    private BigDecimal saldoFinal;
     private List<ValorCategoria> valorCategoria;
 
-    public Resumo(Float vTotalReceitas, Float vTotalDespesas, Float saldoFinal, List<ValorCategoria> valorCategoria) {
+    public Resumo(BigDecimal vTotalReceitas, BigDecimal vTotalDespesas, BigDecimal saldoFinal, List<ValorCategoria> valorCategoria) {
         this.vTotalReceitas = vTotalReceitas;
         this.vTotalDespesas = vTotalDespesas;
         this.saldoFinal = saldoFinal;
