+++ src/main/java/br/com/alura/orcamentoapi/model/ValorCategoria.java
@@ -1,20 +1,13 @@
 package br.com.alura.orcamentoapi.model;
 
-import lombok.Getter;
-import lombok.Setter;
 import org.springframework.beans.factory.annotation.Value;
+
+import java.math.BigDecimal;
 
 public interface ValorCategoria {
 
-//    private CategoriaDespesa categoria;
-//    private Float valor;
-//
-//    public ValorCategoria(CategoriaDespesa categoria, Float valor){
-//        this.categoria = categoria;
-//        this.valor = valor;
-//    }
-
     @Value("#{target.categoria.descricao}")
     String getCategoria();
-    Float getTotal();
+
+    BigDecimal getTotal();
 }
